[refactor] components/modals 폴더를 components/modals/category 폴더로 분리
